requirements.txt
pandas>=2.1.0
requests==2.31.0
numpy>=1.26.0
python-dotenv==1.0.0
tenacity==8.2.3
pytest==7.4.0